CustomerController + adding new Customer form
