+++ eshop-spring-mvc/src/main/resources/Texts.properties
@@ -18,6 +18,9 @@ shopping.show.title=eShop overview
 shopping.product.title=Product {0}
 shopping.category.title=Category {0}
 
+newCustomer.title=New Customer
+newCustomer.submit=submit
+
 #validation error messages
 Size.productCreate.description=The description must be between 3 and 500 characters long.
 Size.productCreate.name=The product name must be between 3 and 50 characters long.
