+++ eshop-spring-mvc/src/main/resources/Texts_cs.properties
@@ -18,3 +18,6 @@ example.redirect.button=Vyvolat p\u0159esm\u011Brov\u00E1n\u00ED
 example.attribute=atribut modelu
 example.value=hodnota
 example.msg_with_params=Po\u017Eadavek byl p\u0159esm\u011Brov\u00E1n {0}. Znovu na\u010Dt\u011Bte str\u00E1nku pro odstran\u011Bn\u00ED t\u00E9to zpr\u00E1vy !
+
+newCustomer.title=Nov\u00fd z\u00e1kazn\u00edk
+newCustomer.submit=odeslat
